+++ src/app/career/page.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function Page() {
+  return <div>career</div>
+}
