+++ src/app/project/page.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function Page() {
+  return <div>project</div>
+}
