+++ src/app/skill/page.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function Page() {
+  return <div>skill</div>
+}
